refactor: restuctured the repository
